feat: adicionar auth context, logout e controle de acesso por role

+++ src/components/Table_components/DeleteButton.jsx
@@ -1,19 +1,29 @@
+import { useAuth } from "../../hooks/useAuth";
 import { Button } from "@mui/material";
 
 export const DeleteButton = ({
   onDelete,
   confirmMessage = "Tem certeza que deseja excluir?",
-}) => (
-  <Button
-    variant="contained"
-    color="error"
-    size="small"
-    onClick={() => {
-      if (window.confirm(confirmMessage)) {
-        onDelete();
-      }
-    }}
-  >
-    Excluir
-  </Button>
-);
+}) => {
+  const { user } = useAuth();
+
+  if (!user || user.role === "editor") {
+    // Esconde o botão para usuários com role editor
+    return null;
+  }
+
+  return (
+    <Button
+      variant="contained"
+      color="error"
+      size="small"
+      onClick={() => {
+        if (window.confirm(confirmMessage)) {
+          onDelete();
+        }
+      }}
+    >
+      Excluir
+    </Button>
+  );
+};
